+++ pyads/utils.py
@@ -9,8 +9,9 @@
 :last modified time: 2018-07-12 14:11:12
 
 """
+import functools
 import sys
-from ctypes import c_ubyte
+import warnings
 
 
 def platform_is_linux():
@@ -24,3 +25,24 @@ def platform_is_windows():
     """Return True if current platform is Windows."""
     # cli being .NET (IronPython)
     return sys.platform == "win32" or sys.platform == "cli"
+
+
+def deprecated(message=None):
+    """Decorator for deprecated functions.
+
+    Shows a deprecation warning with the given message if the
+    decorated function is called.
+
+    """
+    if message is None:
+        message = "Deprecated. This function will not be available in future versions."
+
+    def decorator(func):
+        @functools.wraps(func)
+        def wrapper(*args, **kwargs):
+            warnings.warn(message, DeprecationWarning)
+            return func(*args, **kwargs)
+
+        return wrapper
+
+    return decorator